[change] separate job helper funcs and move to a new helpers folder
