+++ job/job.py
@@ -6,7 +6,7 @@ import subprocess
 import time
 from datetime import datetime, timedelta
 
-from job.helpers import get_site
+from job.helpers.site import get_site
 from job.steps.collaborative_filtering import (
     fetch_data,
     save_defaults,
